Add sidebar and new home page with SRQL query
